+++ packages/plugin-hardhat/src/index.ts
@@ -6,7 +6,7 @@ import { subtask, extendEnvironment, extendConfig } from 'hardhat/config';
 import { TASK_COMPILE_SOLIDITY, TASK_COMPILE_SOLIDITY_COMPILE } from 'hardhat/builtin-tasks/task-names';
 import { lazyObject } from 'hardhat/plugins';
 import { HardhatConfig, HardhatRuntimeEnvironment } from 'hardhat/types';
-import type { silenceWarnings, SolcInput, SolcOutput } from '@openzeppelin/upgrades-core';
+import { assertUnreachable, type silenceWarnings, type SolcInput, type SolcOutput } from '@openzeppelin/upgrades-core';
 import type { DeployFunction } from './deploy-proxy';
 import type { PrepareUpgradeFunction } from './prepare-upgrade';
 import type { UpgradeFunction } from './upgrade-proxy';
@@ -69,7 +69,9 @@ interface RunCompilerArgs {
 }
 
 subtask(TASK_COMPILE_SOLIDITY, async (args: { force: boolean }, hre, runSuper) => {
-  const { readValidations, ValidationsCacheOutdated, ValidationsCacheNotFound } = await import('./utils/validations');
+  const { readValidations, ValidationsCacheOutdated, ValidationsCacheNotFound } = await import(
+    './utils/validations.js'
+  );
 
   try {
     await readValidations(hre);
@@ -85,23 +87,62 @@ subtask(TASK_COMPILE_SOLIDITY, async (args: { force: boolean }, hre, runSuper) =
 });
 
 subtask(TASK_COMPILE_SOLIDITY_COMPILE, async (args: RunCompilerArgs, hre, runSuper) => {
-  const { isNamespaceSupported, validate, solcInputOutputDecoder, makeNamespacedInput } = await import(
-    '@openzeppelin/upgrades-core'
-  );
-  const { writeValidations } = await import('./utils/validations');
+  const { isNamespaceSupported, validate, solcInputOutputDecoder, makeNamespacedInput, trySanitizeNatSpec } =
+    await import('@openzeppelin/upgrades-core');
+  const { writeValidations } = await import('./utils/validations.js');
 
   // TODO: patch input
   const { output, solcBuild } = await runSuper();
 
-  const { isFullSolcOutput } = await import('./utils/is-full-solc-output');
+  const { isFullSolcOutput } = await import('./utils/is-full-solc-output.js');
   if (isFullSolcOutput(output)) {
     const decodeSrc = solcInputOutputDecoder(args.input, output);
 
     let namespacedOutput = undefined;
     if (isNamespaceSupported(args.solcVersion)) {
-      const namespacedInput = makeNamespacedInput(args.input, output);
+      let namespacedInput = makeNamespacedInput(args.input, output, args.solcVersion);
+      namespacedInput = await trySanitizeNatSpec(namespacedInput, args.solcVersion);
       namespacedOutput = (await runSuper({ ...args, quiet: true, input: namespacedInput })).output;
-      await checkNamespacedCompileErrors(namespacedOutput);
+
+      const namespacedCompileErrors = getNamespacedCompileErrors(namespacedOutput);
+      if (namespacedCompileErrors.length > 0) {
+        // If there are compile errors in the namespaced output, show error or warning if needed, then only use the original output
+
+        const msg = `Failed to compile modified contracts for namespaced storage layout validations:\n\n${namespacedCompileErrors.join('\n')}`;
+        const preamble = [
+          'Please report this at https://zpl.in/upgrades/report. If possible, include the source code for the contracts mentioned in the errors above.',
+          'This step allows for advanced storage modifications such as tight varible packing when performing upgrades with namespaced storage layouts.',
+        ];
+
+        switch (hre.config.namespacedCompileErrors) {
+          case undefined:
+          case 'error': {
+            const { UpgradesError } = await import('@openzeppelin/upgrades-core');
+            const details = [
+              ...preamble,
+              'If you are not using namespaced storage, or if you do not anticipate making advanced modifications to namespaces during upgrades,',
+              "you can set namespacedCompileErrors: 'warn' or namespacedCompileErrors: 'ignore' in your hardhat config to convert this to a warning or to ignore this.",
+            ];
+            throw new UpgradesError(msg, () => details.join('\n'));
+          }
+          case 'warn': {
+            const { logWarning } = await import('@openzeppelin/upgrades-core');
+            const details = [
+              ...preamble,
+              'If you are not using namespaced storage, or if you do not anticipate making advanced modifications to namespaces during upgrades,',
+              "you can set namespacedCompileErrors: 'ignore' in your hardhat config to ignore this.",
+            ];
+            logWarning(msg, details);
+            break;
+          }
+          case 'ignore':
+            break;
+          default:
+            assertUnreachable(hre.config.namespacedCompileErrors);
+        }
+
+        namespacedOutput = undefined;
+      }
     }
 
     const validations = validate(output, decodeSrc, args.solcVersion, args.input, namespacedOutput);
@@ -111,11 +152,7 @@ subtask(TASK_COMPILE_SOLIDITY_COMPILE, async (args: RunCompilerArgs, hre, runSup
   return { output, solcBuild };
 });
 
-/**
- * Checks for compile errors in the modified contracts for namespaced storage.
- * If errors are found, throws an error with the compile error messages.
- */
-async function checkNamespacedCompileErrors(namespacedOutput: SolcOutput) {
+function getNamespacedCompileErrors(namespacedOutput: SolcOutput) {
   const errors = [];
   if (namespacedOutput.errors !== undefined) {
     for (const error of namespacedOutput.errors) {
@@ -124,14 +161,7 @@ async function checkNamespacedCompileErrors(namespacedOutput: SolcOutput) {
       }
     }
   }
-  if (errors.length > 0) {
-    const { UpgradesError } = await import('@openzeppelin/upgrades-core');
-    throw new UpgradesError(
-      `Failed to compile modified contracts for namespaced storage:\n\n${errors.join('\n')}`,
-      () =>
-        'Please report this at https://zpl.in/upgrades/report. If possible, include the source code for the contracts mentioned in the errors above.',
-    );
-  }
+  return errors;
 }
 
 extendEnvironment(hre => {
@@ -182,7 +212,7 @@ extendConfig((config: HardhatConfig) => {
 
 if (tryRequire('@nomicfoundation/hardhat-verify')) {
   subtask('verify:etherscan').setAction(async (args, hre, runSuper) => {
-    const { verify } = await import('./verify-proxy');
+    const { verify } = await import('./verify-proxy.js');
     return await verify(args, hre, runSuper);
   });
 }