+++ packages/plugin-hardhat/test/defender-deploy.js
@@ -109,6 +109,8 @@ test('calls defender deploy', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -137,6 +139,8 @@ test('calls defender deploy with relayerId', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -165,6 +169,8 @@ test('calls defender deploy with salt', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -193,6 +199,8 @@ test('calls defender deploy with createFactoryAddress', async t => {
     createFactoryAddress: CREATE_FACTORY,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -221,6 +229,8 @@ test('calls defender deploy with license', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -251,6 +261,8 @@ test('calls defender deploy - licenseType', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -281,6 +293,8 @@ test('calls defender deploy - verifySourceCode false', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -311,6 +325,8 @@ test('calls defender deploy - skipLicenseType', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -392,6 +408,8 @@ test('calls defender deploy - no contract license', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -420,6 +438,8 @@ test('calls defender deploy - unlicensed', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -448,6 +468,8 @@ test('calls defender deploy with constructor args', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -476,6 +498,8 @@ test('calls defender deploy with constructor args with array', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -504,6 +528,8 @@ test('calls defender deploy with verify false', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -532,6 +558,8 @@ test('calls defender deploy with ERC1967Proxy', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 });
 
@@ -558,6 +586,8 @@ test('calls defender deploy with ERC1967Proxy - ignores constructorArgs', async 
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 });
 
@@ -584,6 +614,8 @@ test('calls defender deploy with ERC1967Proxy - ignores empty constructorArgs', 
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 });
 
@@ -610,6 +642,8 @@ test('calls defender deploy with BeaconProxy', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 });
 
@@ -639,6 +673,8 @@ test('calls defender deploy with TransparentUpgradeableProxy', async t => {
     createFactoryAddress: undefined,
     txOverrides: undefined,
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 });
 
@@ -670,6 +706,8 @@ test('calls defender deploy with txOverrides.gasLimit', async t => {
       maxPriorityFeePerGas: undefined,
     },
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -703,6 +741,8 @@ test('calls defender deploy with txOverrides.gasPrice', async t => {
       maxPriorityFeePerGas: undefined,
     },
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -738,6 +778,8 @@ test('calls defender deploy with txOverrides.maxFeePerGas and txOverrides.maxPri
       maxPriorityFeePerGas: '0xa',
     },
     libraries: undefined,
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -772,6 +814,8 @@ test('calls defender deploy with external library', async t => {
     libraries: {
       'contracts/ExternalLibraries.sol:SafeMath': EXTERNAL_LIBRARY_ADDRESS,
     },
+    metadata: undefined,
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);
@@ -808,6 +852,48 @@ test('calls defender deploy with multiple external libraries', async t => {
       'contracts/ExternalLibraries.sol:SafeMath': EXTERNAL_LIBRARY_ADDRESS,
       'contracts/ExternalLibraries.sol:SafeMathV2': EXTERNAL_LIBRARY_2_ADDRESS,
     },
+    metadata: undefined,
+    origin: 'Hardhat',
+  });
+
+  assertResult(t, result);
+});
+
+test('calls defender deploy with metadata', async t => {
+  const { spy, deploy, fakeHre, fakeChainId } = t.context;
+
+  const contractPath = 'contracts/Greeter.sol';
+  const contractName = 'Greeter';
+
+  const factory = await ethers.getContractFactory(contractName);
+  const result = await deploy.defenderDeploy(fakeHre, factory, {
+    metadata: {
+      commitHash: '4ae3e0d',
+      tag: 'v1.0.0',
+      anyOtherField: 'anyValue',
+    },
+  });
+
+  const buildInfo = await hre.artifacts.getBuildInfo(`${contractPath}:${contractName}`);
+  sinon.assert.calledWithExactly(spy, {
+    contractName: contractName,
+    contractPath: contractPath,
+    network: fakeChainId,
+    artifactPayload: JSON.stringify(buildInfo),
+    licenseType: undefined,
+    constructorBytecode: '0x',
+    verifySourceCode: true,
+    relayerId: undefined,
+    salt: undefined,
+    createFactoryAddress: undefined,
+    txOverrides: undefined,
+    libraries: undefined,
+    metadata: {
+      commitHash: '4ae3e0d',
+      tag: 'v1.0.0',
+      anyOtherField: 'anyValue',
+    },
+    origin: 'Hardhat',
   });
 
   assertResult(t, result);